refactor: seperate relationship and relationshipauditlogentry and generate code
