+++ packages/enmeshed_types/lib/src/dtos/dtos.dart
@@ -18,6 +18,7 @@ export 'message_with_attachments.dart';
 export 'peer_deletion_info.dart';
 export 'recipient.dart';
 export 'relationship.dart';
+export 'relationship_audit_log_entry.dart';
 export 'relationship_template.dart';
 export 'request_validation_result.dart';
 export 'setting.dart';
